Add Ai::Catalog::ItemVersionDependency model and checks

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/201585

Merged-by: Keeyan Nejad <knejad@gitlab.com>
Approved-by: Jarka Košanová <jarka@gitlab.com>

+++ ee/app/services/ai/catalog/items/base_destroy_service.rb
@@ -42,7 +42,7 @@ module Ai
         end
 
         def delete_item
-          return item.soft_delete if item.consumers.any?
+          return item.soft_delete if item.consumers.any? || item.reverse_dependencies.any?
 
           item.destroy
         end
